LPS-196245 Delete DDMFormWebConfigurationActivator and move configuration to its usages
