--- modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-web/src/main/java/com/liferay/dynamic/data/mapping/form/web/internal/configuration/activator/DDMFormWebConfigurationActivator.java
@@ -1,39 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.dynamic.data.mapping.form.web.internal.configuration.activator;
-
-import com.liferay.dynamic.data.mapping.form.web.internal.configuration.DDMFormWebConfiguration;
-import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
-
-import java.util.Map;
-
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Modified;
-
-/**
- * @author Leonardo Barros
- */
-@Component(
-	configurationPid = "com.liferay.dynamic.data.mapping.form.web.internal.configuration.DDMFormWebConfiguration",
-	service = DDMFormWebConfigurationActivator.class
-)
-public class DDMFormWebConfigurationActivator {
-
-	public DDMFormWebConfiguration getDDMFormWebConfiguration() {
-		return _ddmFormWebConfiguration;
-	}
-
-	@Activate
-	@Modified
-	protected void activate(Map<String, Object> properties) {
-		_ddmFormWebConfiguration = ConfigurableUtil.createConfigurable(
-			DDMFormWebConfiguration.class, properties);
-	}
-
-	private volatile DDMFormWebConfiguration _ddmFormWebConfiguration;
-
-}
